Move partitioning of search results into separate class
